+++ R/00_utils.R
@@ -108,16 +108,4 @@ with_openxml_quotes <- function(x) {
 
   paste("\n\n``````{=openxml}", x, "``````\n\n", sep = "\n")
 }
-with_openxml_unquote <- function(x) {
-  if (length(x) < 1) {
-    x <- ""
-  } else if (length(x) > 1) {
-    x <- paste(x, collapse = "\n")
-  }
 
-  paste("\n\n``````{=openxml}\n", x, "\n``````\n\n", sep = "\n")
-}
-
-with_html_quotes <- function(x) {
-  paste("```{=html}", x, "```\n\n", sep = "\n")
-}