+++ NEWS.md
@@ -6,6 +6,7 @@
 - doc inconsistency for `set_table_properties()` with layout that defaults to 
 "fixed".
 - add_header_row produced an error after using `delete_column()` (#676)
+- `fmt_signif_after_zeros()` fixed issue with 0 rounding
 
 # flextable 0.9.7